requirements.txt
django==3.2
djangorestframework
psycopg2-binary
pillow